Some Update

+++ Milestone 1/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: abnsila <abnsila@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 16:48:20 by abnsila           #+#    #+#             */
-/*   Updated: 2024/12/06 11:53:21 by abnsila          ###   ########.fr       */
+/*   Updated: 2024/12/06 15:57:59 by abnsila          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int ft_print_char(int c);
 int ft_print_str(char *s);
 int ft_print_nbr(int n);
 int ft_print_hexa(int n, char *base);
-int ft_print_ptr(unsigned int n);
+int ft_print_ptr(unsigned long n);
 int ft_print_unsigned_nbr(unsigned int n);
 
 #endif
